Rename ugly detail file
